safe but infallible extend (wip)
This is a work in progress idea I had to use heap.first.size to cache pending
extend requests that were too small to safely execute. Once there *is* enough
space to align and store a Hole, we can do so.

This does introduce possible "dead_space" between the last allocation and the
newly created Hole after extending, due to the need to align upward before
writing the new Hole. The dead space may be reclaimed sometimes, but perhaps
it's impossible to guarantee that. So this idea might be bad, though it only
affects odd-aligned heap.top() situations.

When we dealloc, would it be sound to *align_up_size* the Layout being free'd,
knowing that we only ever put Holes at aligned addresses?

Example (assume 64-bit arch):
1. New heap: aligned bottom, but size 17
2. Alloc is requested of size 17, align 1, and succeeds. No free holes exist.
3. Heap extend by 15 (to align for Hole) + 16 (Hole size). Extend succeeds.
4. Heap looks like this: [17b in use][15b dead][16b Hole]
5. First allocation is free'd, and the Hole size is rounded up
   Note: we need to check against heap.size() and heap.pending_extend() before
   rounding up all the way
6. Heap looks like this: [32b Hole][16b Hole]